chore: move loading reducer
