+++ packages/common/store/reducers/index.js
@@ -9,5 +9,5 @@ export { default as appUpdates } from '@update/store/reducers/appUpdates';
 export { default as voting } from '@dpos/validator/store/reducers/voting';
 export { default as watchList } from '@dpos/validator/store/reducers/watchList';
 export { default as service } from './service';
-export { default as loading } from './loading';
+export { default as loading } from 'src/modules/common/store/reducers/loading';
 export { account } from 'src/modules/account/store/reducer';
