+++ src/modules/common/store/reducers/loading.js
